pyproject.toml
# ==============================================================================
# pyproject.toml - 现代Python项目配置
#
# 面试要点：
# "我用pyproject.toml统一管理项目元数据、依赖和工具配置，
#  这是PEP 517/518推荐的现代标准，比setup.py更清晰。"
# ==============================================================================

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "hermes-agent-cn"
version = "0.1.0"
description = "生产级Hermes-inspired AI Agent，基于阿里云百炼(Dashscope)构建"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "GaoFeng", email = "2082703545@qq.com"}
]
keywords = ["ai", "agent", "llm", "qwen", "dashscope", "planning"]

dependencies = [
    # LLM接口 - 阿里云百炼(OpenAI兼容模式，无需额外SDK)
    "openai>=1.30.0",          # 用OpenAI SDK调Dashscope兼容接口

    # 新增：
    "chromadb>=1.0.0",      # 向量数据库（1.0+版本Windows兼容稳定）
    "numpy>=1.24.0",        # chromadb间接依赖，显式声明

    # 数据存储
    "sqlite-utils>=3.35",      # SQLite操作增强库

    # 终端界面
    "rich>=13.0.0",            # 彩色终端输出、进度条、表格
    "questionary>=2.0.0",      # 交互式CLI（选择框、输入框）

    # 工具类
    "pyyaml>=6.0",             # YAML配置解析
    "python-dotenv>=1.0.0",    # .env文件加载
    "colorama>=0.4.6",         # Windows CMD颜色支持（rich的底层依赖）
    "tqdm>=4.65.0",            # 进度条（备用）
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
# 安装后可直接在命令行运行 `hermes-cn`
hermes-cn = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

# ==============================================================================
# 工具配置
# ==============================================================================

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N"]
ignore = ["E501"]  # 长行由black处理

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true  # 部分第三方库无stubs

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --cov=src --cov-report=term-missing"